Move LocalDependencyResolver out to separate file, share with direct actor transport
